Extract buttons
