+++ app/components/enterprise_edition/banner_component.html.erb
@@ -23,12 +23,8 @@
           end
         end
 
-        flex.with_row(mt: 2, flex_layout: true, align_items: :center) do |button_layout|
-          buttons.each_with_index do |button, i|
-            button_layout.with_column(ml: (i == 0 ? 0 : 2)) do
-              button
-            end
-          end
+        flex.with_row(mt: 2) do |button_layout|
+          render EnterpriseEdition::UpsaleButtonsComponent.new(feature_key)
         end
       end
     end
